OCR/OmniParser/requirements.txt
bezier
opencv-python
Pillow
pycocotools
tqdm
editdistance
shapely
timm
pandas
transformers